Move customer reports to a separate class
