+++ src/main/java/com/abc/Bank.java
@@ -1,10 +1,11 @@
 package com.abc;
 
 import java.util.ArrayList;
-import java.util.List;
+import java.util.Collection;
+import java.util.Collections;
 
 public class Bank {
-    private final List<Customer> customers;
+    private final Collection<Customer> customers;
 
     public Bank() {
         customers = new ArrayList<Customer>();
@@ -14,22 +15,8 @@ public class Bank {
         customers.add(customer);
     }
 
-    public String customerSummary() {
-        StringBuilder summary = new StringBuilder("Customer Summary");
-        for (Customer customer : customers) {
-            summary.append("\n - ")
-                    .append(customer.getName())
-                    .append(" (")
-                    .append(format(customer.getAccounts().size(), "account"))
-                    .append(')');
-        }
-        return summary.toString();
-    }
-
-    //Make sure correct plural of word is created based on the number passed in:
-    //If number passed in is 1 just return the word otherwise add an 's' at the end
-    private static String format(int number, String word) {
-        return number + " " + (number == 1 ? word : word + "s");
+    public Collection<Customer> getCustomers() {
+        return Collections.unmodifiableCollection(customers);
     }
 
     public double totalInterestPaid() {
